💰 Conversion des prix en entiers sans décimales
- Base de données : DECIMAL(10,2) → INT(11) pour price, balance, amount, proposed_price
- Backend : parseFloat() → parseInt() dans transactions.js
- Frontend : suppression de tous les .toFixed(2) pour l'affichage des prix
- Migration SQL : conversion automatique avec préservation des données existantes

Affichage uniforme : "42 radis" au lieu de "42.00 radis"

+++ backend/routes/transactions.js
@@ -24,9 +24,9 @@ router.get('/', authenticateToken, async (req, res) => {
     let runningBalance = 0;
     const transactionsWithBalance = transactions.map(transaction => {
       if (transaction.to_user_id === req.user.id) {
-        runningBalance += parseFloat(transaction.amount);
+        runningBalance += parseInt(transaction.amount);
       } else if (transaction.from_user_id === req.user.id) {
-        runningBalance -= parseFloat(transaction.amount);
+        runningBalance -= parseInt(transaction.amount);
       }
       
       return {
